Nearest Geographical Neighbours (Discoidal)

Return the neighbours of a language residing within a disc of a specified radius from that language.

nearest_geo_neighbours_discoidal(id, radius, data = WALS)

Arguments

id
Language's WALS ID

radius
Radius of disc in km

data
Dataset to consider; by default the entire WALS

Details

Computed using great_circle_distance. By default (data = WALS), all languages in the WALS atlas will be considered in the computation of the nearest neighbours. By supplying a non-default data argument, it is possible to override this behaviour, e.g. to supply only a subset of the languages.

Examples

# Neighbours of Finnish within 1000 km:
nearest_geo_neighbours_discoidal("fin", radius=1000)

#>      language_ID               language  distance
#> 2045         fsl  Finnish Sign Language  52.20234
#> 896          est               Estonian 338.03809
#> 2522         swv Swedish (Västerbotten) 363.42974
#> 1709         vot                  Votic 388.48079
#> 2422         izh                  Izhor 398.89866
#> 2108         krl               Karelian 417.18734
#> 512          scs  Saami (Central-South) 506.75511
#> 500          lat                Latvian 558.81411
#> 2111         liv                    Liv 577.26795
#> 2421         vep                   Veps 582.39762
#> 781          swe                Swedish 582.39762
#> 2038         svt  Svenska Teckenspråket 640.15912
#> 2164         rnr Romani (North Russian) 733.78682
#> 2464         ylt   Yiddish (Lithuanian) 778.36449
#> 1178         sno       Saami (Northern) 779.70291
#> 563          lit             Lithuanian 780.50557
#> 2329         ski         Saami (Kildin) 796.96301
#> 1980         krm                 Karaim 815.44380
#> 2173         nte        Norsk Tegnspråk 853.93985
#> 214          nor              Norwegian 906.14215
#> 1453         blr            Belorussian 906.70727
#> 2425         ksu              Kashubian 979.23099
#> 495          rus                Russian 996.46686

# Neighbours of Finnish within 1000 km, within the Uralic family:
nearest_geo_neighbours_discoidal("fin", radius=1000,
  data=filter_WALS(family=="Uralic"))

#>    language_ID              language distance
#> 12         est              Estonian 338.0381
#> 18         vot                 Votic 388.4808
#> 27         izh                 Izhor 398.8987
#> 22         krl              Karelian 417.1873
#> 9          scs Saami (Central-South) 506.7551
#> 23         liv                   Liv 577.2680
#> 26         vep                  Veps 582.3976
#> 16         sno      Saami (Northern) 779.7029
#> 25         ski        Saami (Kildin) 796.9630